A testing program for object recognition
It's a simple testing programme using tensorflow based on CIFAR 10 database. It can detect objects in ".png" format pictures in any size and resolution. 
The programme along with the test pictures are in this repository. To use this programme, please download the database file from https://www.cs.toronto.edu/~kriz/cifar.html and unzip it in the same directory with the recognition file and the pictures.